Add files via upload
The goal of this project was to modify the RubberLines code to let the user choose 
a different color for each new line drawn. The ColorPicker application needed to be displayed at 
the top left corner of the JavaFX application. Under the Project9_10 class extends Application I 
initialize the ColorPicker as “colorpicker”; I set it to private. Then I made the colorpicker value 
to be cyan by default under the start function. Under the processMousePress event handler I 
set the “currentLine” stroke color to be the value of the “colorpicker”.